Add basic functionality to symptom checker
